Fixed warnings

+++ src/game/GamePanel.java
@@ -51,7 +51,7 @@ public class GamePanel extends JPanel implements Runnable {
     @Override
     public void run() {
 
-        double drawInterval = 1000000000 / FPS;
+        double drawInterval = (double) 1000000000 / FPS;
         double delta = 0;
         double lastTime = System.nanoTime();
         long currentTime;
